Add tables (t_campaigns) and implementation Company class.

+++ src/main/java/com/example/REST/controller/AdvertiserController.java
@@ -7,8 +7,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 import java.util.Map;
 
 @Controller
@@ -40,7 +38,7 @@ public class AdvertiserController {
 
     @PostMapping("/news/save")
     public String saveNewAdvertiser(@ModelAttribute("customer") Advertiser advertiser) {
-        advertiser.setId(advertiserService.listAll().size() + 1);
+        advertiser.setId((long) (advertiserService.listAll().size() + 1));
         advertiserService.save(advertiser);
         return "redirect:/news";
     }
